Bug Fix- Shadow, Reflectance

+++ parser.h
@@ -17,7 +17,7 @@ namespace parser
 
     struct Vec4f
     {
-        float x, y, z, w;
+        double x, y, z, w;
     };
 
     struct Camera
@@ -26,7 +26,7 @@ namespace parser
         Vec3f gaze;
         Vec3f up;
         Vec4f near_plane;
-        float near_distance;
+        double near_distance;
         int image_width, image_height;
         std::string image_name;
     };
@@ -44,7 +44,7 @@ namespace parser
         Vec3f diffuse;
         Vec3f specular;
         Vec3f mirror;
-        float phong_exponent;
+        double phong_exponent;
     };
 
     struct Face
@@ -70,14 +70,14 @@ namespace parser
     {
         int material_id;
         int center_vertex_id;
-        float radius;
+        double radius;
     };
 
     struct Scene
     {
         //Data
         Vec3i background_color;
-        float shadow_ray_epsilon;
+        double shadow_ray_epsilon;
         int max_recursion_depth;
         std::vector<Camera> cameras;
         Vec3f ambient_light;
